app/requirements.txt
hazelcast-python-client==5.0.1